% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmbryoSelection.R
\name{risk_prediction_analytical}
\alias{risk_prediction_analytical}
\title{Unified Risk Prediction}
\usage{
risk_prediction_analytical(
  r2,
  K,
  n,
  selection_strategy = c("lowest_prs", "exclude_percentile"),
  exclusion_q = NULL,
  qf = NULL,
  qm = NULL,
  random_strategy = c("Fixed", "Binomial", "Poisson"),
  p_lb = 1
)
}
\arguments{
\item{r2}{PRS accuracy (0-1).}

\item{K}{Disease prevalence (0-1).}

\item{n}{Number of embryos (Fixed) or N parameter (Binomial/Poisson).}

\item{selection_strategy}{"lowest_prs" or "exclude_percentile".}

\item{exclusion_q}{Percentile threshold for exclusion (required if strategy is exclude).}

\item{qf}{Father's PRS quantile (0-1). Optional.}

\item{qm}{Mother's PRS quantile (0-1). Optional.}

\item{random_strategy}{"Fixed", "Binomial", or "Poisson".}

\item{p_lb}{Probability of live birth (required for Binomial/Poisson).}
}
\value{
A list: baseline (random risk), selection (strategy risk), rr (relative reduction).
}
\description{
Calculates disease risk and reduction using numerical integration.
Supports population-level or parental-conditional estimates across all strategies.
}
\examples{
# Expected when selecting the minimum PRS embryo
risk_prediction_analytical(0.05, 0.01, 5)
# Excluding the top 5\% PRS
risk_prediction_analytical(0.05, 0.01, 5, 
selection_strategy = "exclude", 
exclusion_q = 0.05)
# The risk when one parent is in the top 5\% of PRS, and the second in the median.
risk_prediction_analytical(0.05, 0.01, 5, 
selection_strategy = "lowest", 
qf = 0.95, qm = 0.5)
# The risk when one parent is in the top 5\% of PRS, and the second in the median.
# and excluding embryos below the 50\% percentile.
risk_prediction_analytical(0.05, 0.01, 5, 
selection_strategy = "exclude",
exclusion_q = 0.5,
qf = 0.95, qm = 0.5)
Same as the previous, but in the binomial/poisson case.
risk_prediction_analytical(0.05, 0.01, 10, 
random_strategy = "Binomial", 
p_lb = 0.5)
risk_prediction_analytical(0.05, 0.01, 10, 
random_strategy = "Poisson", 
p_lb = 0.5)
risk_prediction_analytical(0.05, 0.01, 10, qf = 0.95, qm = 0.5,
random_strategy = "Binomial", 
p_lb = 0.5)
risk_prediction_analytical(0.05, 0.01, 10, qf = 0.95, qm = 0.5,
random_strategy = "Poisson", 
p_lb = 0.5)
}
