% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{shiny_calculator}
\alias{shiny_calculator}
\title{Shiny based calculator for risk reductions.}
\usage{
shiny_calculator()
}
\description{
Opens a shiny calculator which allows the user to estimate the expected risk and risk reduction
of PRS based embryo selection.
}
\examples{
shiny_calculator()

}
